Update returnable line items to take into account previous returns #179
